refactor(i18n): simplify resolver modules
